`single?` fn and the sequence abstraction.
I noticed a lot of `(= 1 (count path))` throughout the codebase, and so
introduced a `single?`, if only to reduce the parentheses clutter. Here,
I've only used it on these two forms, but I'll start using it wherever I
can as I move through the codebase.

Second, we can recur on the `rest` of the path, instead of using a
subvec, because we only rely on the sequence abstraction, not on the
paths being vectors.