An example of a type implementing core interfaces.
This is the last thing I want to introduce, because it's pretty neat,
but how you do it is not made clear in any documentation I've found.
This is an example of a silly type `Foo` implementing core interfaces
that are defined in Java. In order to do this, you must use a `deftype`
instead of `defrecord`, which means you lose all of the "treat your
object like a map" goodness, but you gain a bit of power.

This `Foo` type will respond to numerous built-in Clojure functions: I
can map, filter, or reduce over it; I can conj onto it; I can disj items
from it (it's assumed that it's instantiated with a set); I can ask if
it contains? anything; I can turn it into a seq; I can empty it; et al.
One function I didn't implement, but that you should, is `(equiv [self
o])`, in the IPersistentCollection interface, that will handle equality.

So, this would be a pretty cool thing for OrderedSet to do -- instead of
having an `oset-insert` and `oset-remove`, you'd just conj and disj on
it like any other set. Cool!